refactor: step 4 content
